Make 100% sure the sector number won't ever overflow 32 bits.
svn:sync-xref-src-repo-rev: r28440

+++ src/VBox/Devices/Storage/VmdkHDDCore.cpp
@@ -1,4 +1,4 @@
-/** $Id: VmdkHDDCore.cpp 7156 2008-02-26 17:59:24Z klaus.espenlaub@oracle.com $ */
+/** $Id: VmdkHDDCore.cpp 7157 2008-02-26 18:01:40Z klaus.espenlaub@oracle.com $ */
 /** @file
  * VMDK Disk image, Core Code.
  */
@@ -569,7 +569,7 @@ static int vmdkCreateGrainDirectory(PVMDKEXTENT pExtent, uint64_t uStartSector,
     if (fPreAlloc)
     {
         uint32_t uGTSectorLE;
-        uint32_t uOffsetSectors;
+        uint64_t uOffsetSectors;
 
         uOffsetSectors = pExtent->uSectorRGD + VMDK_BYTE2SECTOR(cbGDRounded);
         for (i = 0; i < pExtent->cGDEntries; i++)
